+++ client/mobile/ios/rac/rac/Interactive/Text/ChatMessagesView.swift
@@ -12,9 +12,10 @@ enum ChatRole {
 }
 
 struct ChatMessage: Identifiable {
-    let id: UUID
+    var id: String
     let role: ChatRole
     var content: String
+    var finalized: Bool = false
 }
 
 struct ChatMessagesView: View {
@@ -35,7 +36,7 @@ struct ChatMessagesView: View {
                 ForEach(messages) { message in
                     switch message.role {
                     case .assistant:
-                        CharacterMessage(message: message.content)
+                        CharacterMessage(message: message)
                             .listRowSeparator(.hidden)
                             .listRowBackground(Constants.realBlack)
                             .id(messages.firstIndex(where: { $0.id == message.id}))
@@ -93,154 +94,13 @@ struct ChatMessagesView: View {
 struct ChatMessagesView_Previews: PreviewProvider {
     static var previews: some View {
         ChatMessagesView(messages: .constant([
-            ChatMessage(id: UUID(), role: .assistant, content: "Hello stranger, what’s your name?"),
-            ChatMessage(id: UUID(), role: .user, content: "Hi 👋 my name is Karina"),
-            ChatMessage(id: UUID(), role: .assistant, content: "Greetings, Karina. What can I do for you?"),
-            ChatMessage(id: UUID(), role: .user, content: "What’s your name?"),
-            ChatMessage(id: UUID(), role: .assistant, content: "I have no name. I am Realtime’s AI soul. I exist in the digital, but if I had to have a name, I would pick Ray 😉"),
-            ChatMessage(id: UUID(), role: .user, content: "Ray is a nice name!"),
-            ChatMessage(id: UUID(), role: .assistant, content: "Well thank you, Karina! I like your nam too. Now tell me, where do you live?")
+            ChatMessage(id: UUID().uuidString, role: .assistant, content: "Hello stranger, what’s your name?"),
+            ChatMessage(id: UUID().uuidString, role: .user, content: "Hi 👋 my name is Karina"),
+            ChatMessage(id: UUID().uuidString, role: .assistant, content: "Greetings, Karina. What can I do for you?"),
+            ChatMessage(id: UUID().uuidString, role: .user, content: "What’s your name?"),
+            ChatMessage(id: UUID().uuidString, role: .assistant, content: "I have no name. I am Realtime’s AI soul. I exist in the digital, but if I had to have a name, I would pick Ray 😉"),
+            ChatMessage(id: UUID().uuidString, role: .user, content: "Ray is a nice name!"),
+            ChatMessage(id: UUID().uuidString, role: .assistant, content: "Well thank you, Karina! I like your nam too. Now tell me, where do you live?")
         ]), isExpectingUserInput: .constant(true), onSendUserMessage: { _ in })
     }
 }
-
-struct CharacterMessage: View {
-    let message: String
-    @State var isShowingCommentSheet = false
-    @State var thumbsUp = false
-    @State var feedback: String = ""
-
-    var body: some View {
-        Text(message)
-          .font(Font.custom("Prompt", size: 20))
-          .foregroundColor(.white)
-          .frame(maxWidth: .infinity, alignment: .topLeading)
-          .contextMenu {
-              Button {
-                  thumbsUp = true
-                  isShowingCommentSheet = true
-              } label: {
-                  Label("Great response!", systemImage: "hand.thumbsup")
-              }
-
-              Button {
-                  thumbsUp = false
-                  isShowingCommentSheet = true
-              } label: {
-                  Label("Something's wrong", systemImage: "hand.thumbsdown")
-              }
-
-              Divider()
-
-              Button {
-                  UIPasteboard.general.string = message
-              } label: {
-                  Label("Copy to Clipboard", systemImage: "doc.on.doc")
-              }
-          }
-          .alert(thumbsUp
-                 ? "Thanks for the feedback! Share more comments?"
-                 : "Sorry to hear that! Share more comments?",
-                 isPresented: $isShowingCommentSheet) {
-              // Comment sheet
-              VStack {
-                  TextField("(Optional) Comments", text: $feedback)
-                      .tint(.black)
-                      .foregroundColor(.black)
-
-                  Button("Cancel") {
-                      isShowingCommentSheet = false
-                  }
-
-                  Button("Submit") {
-                      // Handle submit action along with comment
-                      print("Submitted with comment: \(thumbsUp) \(feedback)")
-                      feedback = ""
-                      isShowingCommentSheet = false
-                  }
-              }
-              .padding()
-          }
-    }
-}
-
-struct UserMessage: View {
-    let message: String
-    var onCancel: (() -> Void)? = nil
-
-    var body: some View {
-        HStack(spacing: 8) {
-            Text(message)
-                .font(Font.custom("Prompt", size: 20))
-                .multilineTextAlignment(.trailing)
-                .foregroundColor(.white)
-                .padding(.vertical, 11)
-                .padding(.leading, 20)
-                .if(onCancel == nil) { view in
-                    view.padding(.trailing, 20)
-                }
-
-            if let onCancel {
-                Button {
-                    onCancel()
-                } label: {
-                    Image(systemName: "xmark.circle")
-                        .padding(12)
-                }
-                .buttonStyle(CustomButtonStyle())
-            }
-        }
-        .background(Color(red: 0.4, green: 0.52, blue: 0.83).opacity(0.25))
-        .roundedCorner(20, corners: [.bottomLeft, .topLeft, .topRight])
-        .frame(maxWidth: .infinity, alignment: .topTrailing)
-    }
-}
-
-struct UserInputView: View {
-    @Binding var message: String
-
-    var body: some View {
-        TextField("Your turn", text: $message, axis: .vertical)
-            .textFieldStyle(CustomTextFieldStyle())
-    }
-}
-
-struct CustomTextFieldStyle: TextFieldStyle {
-    func _body(configuration: TextField<Self._Label>) -> some View {
-        configuration
-            .font(Font.custom("Prompt", size: 20))
-            .multilineTextAlignment(.trailing)
-            .lineLimit(...3)
-            .foregroundColor(.white)
-            .padding(.horizontal, 20)
-            .padding(.vertical, 11)
-            .background(Color(red: 0.4, green: 0.52, blue: 0.83).opacity(0.25))
-            .roundedCorner(20, corners: [.bottomLeft, .topLeft, .topRight])
-    }
-}
-
-extension View {
-    func roundedCorner(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape(RoundedCorner(radius: radius, corners: corners) )
-    }
-
-    func onEnter(@Binding of text: String, action: @escaping () -> ()) -> some View {
-        onChange(of: text) { newValue in
-            if let last = newValue.last, last == "\n" {
-                text.removeLast()
-                action()
-            }
-        }
-    }
-}
-
-struct RoundedCorner: Shape {
-    var radius: CGFloat = .infinity
-    var corners: UIRectCorner = .allCorners
-
-    func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
-
